Initial commit.
This is what I got after an hour and a half, and a 30% battery drain;
rushing a Java prototype interpreter... misses maybe 99% of everything
that must be done in order to get to the level of the C interpreter.

I named it `pangolin` because pangolins brought us COVID-19 so now
everybody is hating them so much.